Changed priorities to longs to cope with double conversation
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@7703 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 3451c5ed8d6b52a305ff591c14aaf46d357ea809

+++ core/base/src/de/uniol/inf/is/odysseus/scheduler/strategy/CurrentPlanPriorityComperator.java
@@ -6,8 +6,8 @@ public class CurrentPlanPriorityComperator implements Comparator<IScheduling> {
 
 	@Override
 	public int compare(IScheduling p1, IScheduling p2) {
-		int p1Priority = p1.getPlan().getCurrentPriority();
-		int p2priority = p2.getPlan().getCurrentPriority();
+		long p1Priority = p1.getPlan().getCurrentPriority();
+		long p2priority = p2.getPlan().getCurrentPriority();
 		if (p1Priority < p2priority){
 			return 1;
 		}
